CODIGO/requirements.txt
python-fasthtml
passlib
bcrypt


